# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TaleMotion Backend Foundation API
  version: 0.1.0
  description: PostgreSQL-backed project, scene, and job foundation.
servers:
  - url: http://localhost:8000/api/v1
security:
  - cookieAuth: []
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - {name: status, in: query, schema: {$ref: "#/components/schemas/ProjectStatus"}}
        - {name: mode, in: query, schema: {$ref: "#/components/schemas/VideoMode"}}
        - {name: search, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, schema: {type: [string, "null"]}}
      responses:
        "200":
          description: Cursor page
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProjectListResponse"}
        "400": {$ref: "#/components/responses/Error"}
    post:
      tags: [Projects]
      summary: Create a project and default Main chapter
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateProjectRequest"}
      responses:
        "201":
          description: Persisted project
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProjectResponse"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /projects/{project_id}:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
    get:
      tags: [Projects]
      summary: Get project with chapters and scenes
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProjectResponse"}
        "404": {$ref: "#/components/responses/Error"}
    patch:
      tags: [Projects]
      summary: Update safe user-controlled fields
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateProjectRequest"}
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProjectResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
    delete:
      tags: [Projects]
      summary: Soft-delete a project
      responses:
        "204": {description: Project soft-deleted}
        "404": {$ref: "#/components/responses/Error"}
  /projects/{project_id}/storyboard:
    post:
      tags: [Projects]
      summary: Queue a four-scene historical storyboard
      parameters:
        - {$ref: "#/components/parameters/ProjectId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateStoryboardRequest"}
      responses:
        "202":
          description: Persisted queued storyboard job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /projects/{project_id}/generations:
    post:
      tags: [Projects]
      summary: Queue media generation for all four storyboard scenes
      parameters:
        - {$ref: "#/components/parameters/ProjectId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateProjectGenerationRequest"}
      responses:
        "202":
          description: Parent job with four queued child summaries
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /projects/{project_id}/renders:
    get:
      tags: [Renders]
      summary: List project render versions
      parameters:
        - {$ref: "#/components/parameters/ProjectId"}
      responses:
        "200":
          description: Render versions, newest first
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RenderListResponse"}
        "404": {$ref: "#/components/responses/Error"}
    post:
      tags: [Renders]
      summary: Queue final MP4 rendering
      parameters:
        - {$ref: "#/components/parameters/ProjectId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateRenderRequest"}
      responses:
        "202":
          description: Persisted queued render job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
  /renders/{render_id}:
    get:
      tags: [Renders]
      summary: Get a persisted render
      parameters:
        - {name: render_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Render metadata
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RenderResponse"}
        "404": {$ref: "#/components/responses/Error"}
  /renders/{render_id}/preview-url:
    post:
      tags: [Renders]
      summary: Create a signed final-video preview URL
      parameters:
        - {name: render_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Signed final-video preview
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SignedPreviewUrlResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: talemotion_session
  parameters:
    ProjectId: {name: project_id, in: path, required: true, schema: {type: string}}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 1, maxLength: 128}
...
  responses:
    Error:
      description: Consistent error envelope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
  schemas:
...
    VideoMode:
      type: string
      enum: [historical_documentary, microdrama, product_advertisement]
    ProjectStatus:
      type: string
      enum: [draft, storyboard_pending, storyboard_generating, storyboard_ready, media_generating, rendering, ready, failed, deleted]
    ChapterStatus:
      type: string
      enum: [draft, ready, generating, completed, failed]
    SceneStatus:
      type: string
      enum: [draft, ready, queued, generating, completed, failed]
    JobType:
      type: string
      enum: [storyboard, scene_generation, scene_regeneration, project_generation, render]
    JobStatus:
      type: string
      enum: [queued, running, completed, failed, cancel_requested, cancelled]
...
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [mode, title, topic, duration_seconds, aspect_ratio]
      properties:
        mode: {$ref: "#/components/schemas/VideoMode"}
        title: {type: string, minLength: 1}
        topic: {type: string, minLength: 1}
        additional_direction: {type: string, default: ""}
        language: {type: string, default: en}
        duration_seconds: {type: integer, enum: [30, 45]}
        aspect_ratio: {type: string, enum: ["9:16"]}
        visual_style: {type: string}
        narration_style: {type: string}
        captions_enabled: {type: boolean, default: true}
        narration_enabled: {type: boolean, default: true}
        music_enabled: {type: boolean, default: true}
        historical_accuracy_note: {type: [string, "null"]}
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      properties:
        title: {type: string, minLength: 1}
        topic: {type: string, minLength: 1}
        additional_direction: {type: string}
        language: {type: string}
        duration_seconds: {type: integer, enum: [30, 45]}
        aspect_ratio: {type: string, enum: ["9:16"]}
        visual_style: {type: string}
        narration_style: {type: string}
        captions_enabled: {type: boolean}
        narration_enabled: {type: boolean}
        music_enabled: {type: boolean}
        historical_accuracy_note: {type: [string, "null"]}
    ProjectResponse:
      allOf:
        - {$ref: "#/components/schemas/CreateProjectRequest"}
        - type: object
          required: [id, status, generation_progress, chapters, created_at, updated_at]
          properties:
            id: {type: string}
            status: {$ref: "#/components/schemas/ProjectStatus"}
            generation_progress: {type: integer, minimum: 0, maximum: 100}
            chapters:
              type: array
              items: {$ref: "#/components/schemas/ChapterResponse"}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
    ProjectListResponse:
      type: object
      required: [items, next_cursor, has_more]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/ProjectResponse"}}
        next_cursor: {type: [string, "null"]}
        has_more: {type: boolean}
    ChapterResponse:
      type: object
      required: [id, project_id, title, position, status, scenes, created_at, updated_at]
      properties:
        id: {type: string}
        project_id: {type: string}
        title: {type: string}
        summary: {type: [string, "null"]}
        position: {type: integer, minimum: 1}
        target_duration_seconds: {type: [integer, "null"]}
        status: {$ref: "#/components/schemas/ChapterStatus"}
        scenes: {type: array, items: {$ref: "#/components/schemas/SceneResponse"}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateSceneRequest:
      type: object
      additionalProperties: false
      required: [title, narration, visual_prompt, duration_seconds]
      properties:
        title: {type: string, minLength: 1}
        narration: {type: string}
        visual_prompt: {type: string}
        duration_seconds: {type: integer, minimum: 1, maximum: 60}
        position: {type: integer, minimum: 1}
...
    CreateStoryboardRequest:
      type: object
      additionalProperties: false
      properties:
        replace_existing: {type: boolean, default: false}
    CreateProjectGenerationRequest:
      type: object
      additionalProperties: false
      properties:
        generate_video: {type: boolean, default: true}
    CreateRenderRequest:
      type: object
      additionalProperties: false
      properties:
        narration_enabled: {type: [boolean, "null"]}
        captions_enabled: {type: [boolean, "null"]}
        music_enabled: {type: [boolean, "null"]}
...
    SceneResponse:
      allOf:
        - {$ref: "#/components/schemas/CreateSceneRequest"}
        - type: object
          required: [id, chapter_id, position, status, active_asset_version, created_at, updated_at]
          properties:
            id: {type: string}
            chapter_id: {type: string}
            status: {$ref: "#/components/schemas/SceneStatus"}
            active_asset_id: {type: [string, "null"]}
            active_asset_version: {type: integer}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
    GenerationJobResponse:
      type: object
      required: [id, project_id, type, status, progress, input_payload, retry_count, max_retries, created_at, updated_at, children]
      properties:
        id: {type: string}
        project_id: {type: string}
        scene_id: {type: [string, "null"]}
        parent_job_id: {type: [string, "null"]}
        type: {$ref: "#/components/schemas/JobType"}
        status: {$ref: "#/components/schemas/JobStatus"}
        progress: {type: integer, minimum: 0, maximum: 100}
        current_stage: {type: [string, "null"]}
        input_payload: {type: object}
        result_payload: {type: [object, "null"]}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
        retry_count: {type: integer}
        max_retries: {type: integer}
        cancel_requested_at: {type: [string, "null"], format: date-time}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        updated_at: {type: string, format: date-time}
        children:
          type: array
          items: {$ref: "#/components/schemas/GenerationJobChildResponse"}
...
    GenerationJobChildResponse:
      type: object
      required: [id, scene_id, status, progress, result_asset_id, error_code, error_message]
      properties:
        id: {type: string}
        scene_id: {type: [string, "null"]}
        status: {$ref: "#/components/schemas/JobStatus"}
        progress: {type: integer, minimum: 0, maximum: 100}
        result_asset_id: {type: [string, "null"]}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
...
    SignedPreviewUrlResponse:
      type: object
      required: [url, expires_at]
      properties:
        url: {type: string, format: uri}
        expires_at: {type: string, format: date-time}
    RenderResponse:
      type: object
      required:
        - id
        - project_id
        - version
        - status
        - narration_enabled
        - captions_enabled
        - music_enabled
        - created_at
        - updated_at
      properties:
        id: {type: string}
        project_id: {type: string}
        job_id: {type: [string, "null"]}
        version: {type: integer, minimum: 1}
        status: {type: string, enum: [queued, rendering, completed, failed, cancelled]}
        asset_id: {type: [string, "null"]}
        duration_seconds: {type: [integer, "null"]}
        file_size_bytes: {type: [integer, "null"]}
        narration_enabled: {type: boolean}
        captions_enabled: {type: boolean}
        music_enabled: {type: boolean}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        updated_at: {type: string, format: date-time}
    RenderListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/RenderResponse"}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            details: {type: object}
            request_id: {type: string}